Separate population components for ease of use in documentation
